backend/requirements.txt
flask
flask-cors
pillow
tensorflow
keras
torch
torchvision
transformers
pandas
numpy
matplotlib
scikit-learn
pypdf2